membership_inference_attacks/requirements.txt
matplotlib
numpy
Pillow
pytorch
torchvision
NearPy
